Some further code cleanup

+++ src/leary.h
@@ -22,7 +22,7 @@
 #include "core/types.h"
 #include "core/allocator.h"
 
-#define ARRAY_SIZE(a) sizeof((a)) / sizeof((a)[0])
+#define ARRAY_SIZE(a) (isize)(sizeof((a)) / sizeof((a)[0]))
 
 #define MIN(a, b) (a) < (b) ? (a) : (b)
 #define MAX(a, b) (a) > (b) ? (a) : (b)
